Tests/LibWeb/Text/expected/async-scrolling/scroll-node-index-after-visual-context-change.txt

scroll node index changed: true
